rewrote package with pedroPathing under teamcode to comply with best practices
