+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/constants/FConstants.java
@@ -1,7 +1,9 @@
-package org.firstinspires.ftc.teamcode.constants;
+package org.firstinspires.ftc.teamcode.pedroPathing.constants;
 
-import com.pedropathing.follower.FollowerConstants;
 import com.pedropathing.localization.Localizers;
+import com.pedropathing.follower.FollowerConstants;
+import com.pedropathing.util.CustomFilteredPIDFCoefficients;
+import com.pedropathing.util.CustomPIDFCoefficients;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
 public class FConstants {
